rf: rename `keyring.py` to `keyring_utils.py`
Doing this avoids `keyring.py` shadowing the
`keyring` dependency pacakge.
